backend_server/requirements.txt
# متطلبات Python للمشروع
fastapi==0.109.0
uvicorn==0.27.0
python-telegram-bot==20.7
sqlalchemy==2.0.25
pydantic==2.5.3
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
aiofiles==23.2.1
redis==5.0.1
httpx==0.26.0
openai==1.10.0
python-dotenv==1.0.0
cryptography==41.0.7
